# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillTwin AI Mentor API",
    "description": "Contract and intelligence source-of-truth for the SkillTwin personalized AI mentor platform. Exposes adaptive goals, winding journey roadmaps, evidence-based learner twin state, real-time mentor guidance, and spaced retention retrieval practice.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/goals": {
      "post": {
        "tags": [
          "Goals"
        ],
        "summary": "Create Goal & Initialize Journey Generation",
        "description": "Creates a new learning goal for the authenticated learner.\nInitializes an adaptive journey in PENDING/PROCESSING state and triggers asynchronous\nbackground roadmap generation, returning the goal and tracking status immediately.",
        "operationId": "create_goal_api_v1_goals_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Goals"
        ],
        "summary": "List Authenticated Learner Goals",
        "description": "List all learning goals owned by the authenticated learner.",
        "operationId": "list_goals_api_v1_goals_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GoalResponse"
                  },
                  "title": "Response List Goals Api V1 Goals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{goal_id}": {
      "get": {
        "tags": [
          "Goals"
        ],
        "summary": "Get Goal by ID",
        "description": "Retrieve goal details and live journey generation status.\nStrictly verifies ownership against the authenticated JWT identity.",
        "operationId": "get_goal_api_v1_goals__goal_id__get",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Goals"
        ],
        "summary": "Partially Update Goal",
        "description": "Partially update properties of an existing goal (e.g. daily minutes, target deadline, constraints).\nStrictly verifies ownership against the authenticated JWT identity.",
        "operationId": "patch_goal_api_v1_goals__goal_id__patch",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/concepts/{concept_id}": {
      "get": {
        "tags": [
          "Concepts"
        ],
        "summary": "Get Concept Detail & Live Learner State",
        "description": "Returns concept curriculum definition together with the learner's live Twin state:\n- Multi-factor mastery score, confidence calibration, retention score, and risk score\n- Status, active misconceptions, evidence history, and next scheduled review",
        "operationId": "get_concept_detail_api_v1_concepts__concept_id__get",
        "parameters": [
          {
            "name": "concept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Concept Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConceptDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/concepts/{concept_id}/evidence": {
      "get": {
        "tags": [
          "Concepts"
        ],
        "summary": "Get Concept Evidence History",
        "description": "Returns the chronological history of empirical evidence proofs\n(recalls, practice problems, teach-backs, projects, delayed retrievals) for this concept.",
        "operationId": "get_concept_evidence_api_v1_concepts__concept_id__evidence_get",
        "parameters": [
          {
            "name": "concept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Concept Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvidenceSchema"
                  },
                  "title": "Response Get Concept Evidence Api V1 Concepts  Concept Id  Evidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConceptDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "difficulty_level": {
            "type": "string",
            "title": "Difficulty Level"
          },
          "prerequisites": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prerequisites"
          },
          "mastery": {
            "type": "number",
            "title": "Mastery"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "retention_risk": {
            "$ref": "#/components/schemas/RetentionRisk"
          },
          "status": {
            "$ref": "#/components/schemas/LearnerConceptStatus"
          },
          "evidence_count": {
            "type": "integer",
            "title": "Evidence Count"
          },
          "mastery_score": {
            "type": "number",
            "title": "Mastery Score",
            "default": 0.0
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score",
            "default": 0.0
          },
          "retention_score": {
            "type": "number",
            "title": "Retention Score",
            "default": 100.0
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score",
            "default": 0.0
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "next_review_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Review At"
          },
          "last_evidence_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Evidence At"
          },
          "misconceptions": {
            "items": {
              "$ref": "#/components/schemas/MisconceptionSchema"
            },
            "type": "array",
            "title": "Misconceptions",
            "default": []
          },
          "evidence_history": {
            "items": {
              "$ref": "#/components/schemas/EvidenceSchema"
            },
            "type": "array",
            "title": "Evidence History",
            "default": []
          },
          "mentor_recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mentor Recommendation"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "domain",
          "difficulty_level",
          "prerequisites",
          "mastery",
          "confidence",
          "retention_risk",
          "status",
          "evidence_count"
        ],
        "title": "ConceptDetailResponse"
      },
...
      "EvidenceSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "concept_id": {
            "type": "string",
            "title": "Concept Id"
          },
          "evidence_type": {
            "$ref": "#/components/schemas/EvidenceType"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "feedback": {
            "type": "string",
            "title": "Feedback"
          },
          "reasoning_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Score"
          },
          "transfer_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer Score"
          },
          "misconceptions_detected": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Misconceptions Detected"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "concept_id",
          "evidence_type",
          "score",
          "feedback",
          "created_at"
        ],
        "title": "EvidenceSchema"
      },
      "EvidenceType": {
        "type": "string",
        "enum": [
          "RECALL",
          "PRACTICE",
          "EXPLANATION",
          "TEACH_BACK",
          "PROJECT",
          "ASSESSMENT",
          "DELAYED_RETRIEVAL"
        ],
        "title": "EvidenceType"
      },
...
      "GenerationStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PROCESSING",
          "READY",
          "FAILED"
        ],
        "title": "GenerationStatus"
      },
      "GoalCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Title",
            "example": "Become a Senior Flutter Engineer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "example": "Master architecture, state management, and async performance"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline",
            "example": "2026-12-31"
          },
          "daily_minutes": {
            "type": "integer",
            "maximum": 480.0,
            "minimum": 5.0,
            "title": "Daily Minutes",
            "default": 30,
            "example": 30
          },
          "current_level": {
            "type": "string",
            "title": "Current Level",
            "default": "beginner",
            "example": "intermediate"
          },
          "existing_knowledge": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Existing Knowledge",
            "example": "Comfortable with basic Dart and UI widgets"
          },
          "constraints": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraints",
            "example": [
              "Only weekdays",
              "No videos longer than 15 min"
            ]
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "GoalCreateRequest"
      },
      "GoalResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "daily_minutes": {
            "type": "integer",
            "title": "Daily Minutes"
          },
          "current_level": {
            "type": "string",
            "title": "Current Level"
          },
          "existing_knowledge": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Existing Knowledge"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Constraints"
          },
          "status": {
            "$ref": "#/components/schemas/GoalStatus"
          },
          "active_journey_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Journey Id"
          },
          "generation_status": {
            "$ref": "#/components/schemas/GenerationStatus",
            "default": "PENDING"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "daily_minutes",
          "current_level",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "GoalResponse"
      },
      "GoalStatus": {
        "type": "string",
        "enum": [
          "active",
          "completed",
          "paused",
          "archived"
        ],
        "title": "GoalStatus"
      },
      "GoalUpdateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "daily_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 480.0,
                "minimum": 5.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Minutes"
          },
          "current_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Level"
          },
          "existing_knowledge": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Existing Knowledge"
          },
          "constraints": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraints"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GoalStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "GoalUpdateRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LearnerConceptStatus": {
        "type": "string",
        "enum": [
          "NOT_STARTED",
          "LEARNING",
          "MASTERED",
          "NEEDS_REVIEW"
        ],
        "title": "LearnerConceptStatus"
      },
...
      "MisconceptionSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "concept_id": {
            "type": "string",
            "title": "Concept Id"
          },
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "severity": {
            "$ref": "#/components/schemas/MisconceptionSeverity"
          },
          "status": {
            "$ref": "#/components/schemas/MisconceptionStatus"
          },
          "detected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Detected At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "concept_id",
          "tag",
          "description",
          "severity",
          "status",
          "detected_at"
        ],
        "title": "MisconceptionSchema"
      },
      "MisconceptionSeverity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "MisconceptionSeverity"
      },
      "MisconceptionStatus": {
        "type": "string",
        "enum": [
          "active",
          "resolved"
        ],
        "title": "MisconceptionStatus"
      },
...
      "RetentionRisk": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "RetentionRisk"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}